Bætt við takka sem færir okkur yfir í AllBooksActivity. Allar bækur teknar af forsíðu.
