+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".RequestBookActivity"></activity>
+        <activity android:name=".AllBooksActivity"></activity>
+        <activity android:name=".RequestBookActivity" />
         <activity android:name=".AddBookForSaleActivity" />
         <activity android:name=".RegisterActivity" />
         <activity android:name=".LoginActivity" />
