+++ app/src/main/java/com/example/bokamarkadur/MainActivity.java
@@ -5,23 +5,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import android.util.Log;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.bokamarkadur.POJO.Book;
-import com.example.bokamarkadur.POJO.BookList;
-import com.example.bokamarkadur.POJO.BooksAdapter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -29,7 +14,8 @@ public class MainActivity extends AppCompatActivity {
     private Button registerbutton;
     private Button AddBook;
     private Button RequestBook;
-    private static final String TAG = MainActivity.class.getSimpleName();
+    private Button AllBooksBtn;
+//    private static final String TAG = MainActivity.class.getSimpleName();
 
     APIInterface apiInterface;
     private TextView textViewResult;
@@ -39,42 +25,42 @@ public class MainActivity extends AppCompatActivity {
             setContentView(R.layout.activity_main);
 
 
-            /**
-             * Þórdís start hér
-             */
-            apiInterface = APIClient.getClient().create(APIInterface.class);
-
-            final RecyclerView recyclerView = findViewById(R.id.books_recycler_view);
-            recyclerView.setLayoutManager(new LinearLayoutManager(this));
-            recyclerView.setAdapter(new BooksAdapter(new ArrayList<Book>(), R.layout.list_item, getApplicationContext()));
-
-
-            /**
-             GET kall sem skilar lista af öllum bókum.
-             **/
-            Call<BookList> getAllBooks = apiInterface.getBooks();
-            getAllBooks.enqueue(new Callback<BookList>() {
-                @Override
-                public void onResponse(Call<BookList> call, Response<BookList> response) {
-                    int statusCode = response.code();
-                    List<Book> books = response.body().getBooks();
-                    recyclerView.setAdapter(new BooksAdapter(books, R.layout.list_item, getApplicationContext()));
-
-
-                    Log.d(TAG, "Number of books received: " + books.size());
-
-
-                    // Má eyða - Birtir toast ef svar hefur borist.
-                    Toast.makeText(getApplicationContext(), "Response received", Toast.LENGTH_LONG).show();
-                }
-
-                @Override
-                public void onFailure(Call<BookList> call, Throwable t) {
-                    // Log error here since request failed
-                    Log.e(TAG, t.toString());
-                    call.cancel();
-                }
-            });
+//            /**
+//             * Þórdís start hér
+//             */
+//            apiInterface = APIClient.getClient().create(APIInterface.class);
+//
+//            final RecyclerView recyclerView = findViewById(R.id.books_recycler_view);
+//            recyclerView.setLayoutManager(new LinearLayoutManager(this));
+//            recyclerView.setAdapter(new BooksAdapter(new ArrayList<Book>(), R.layout.list_item, getApplicationContext()));
+//
+//
+//            /**
+//             GET kall sem skilar lista af öllum bókum.
+//             **/
+//            Call<BookList> getAllBooks = apiInterface.getBooks();
+//            getAllBooks.enqueue(new Callback<BookList>() {
+//                @Override
+//                public void onResponse(Call<BookList> call, Response<BookList> response) {
+//                    int statusCode = response.code();
+//                    List<Book> books = response.body().getBooks();
+//                    recyclerView.setAdapter(new BooksAdapter(books, R.layout.list_item, getApplicationContext()));
+//
+//
+//                    Log.d(TAG, "Number of books received: " + books.size());
+//
+//
+//                    // Má eyða - Birtir toast ef svar hefur borist.
+//                    Toast.makeText(getApplicationContext(), "Response received", Toast.LENGTH_LONG).show();
+//                }
+//
+//                @Override
+//                public void onFailure(Call<BookList> call, Throwable t) {
+//                    // Log error here since request failed
+//                    Log.e(TAG, t.toString());
+//                    call.cancel();
+//                }
+//            });
 
 
             //add book 4 sale
@@ -111,6 +97,14 @@ public class MainActivity extends AppCompatActivity {
                     openRegisterActivity();
                 }
             });
+            // Förum yfir í AllBooksActivity þar sem allar bækur eru birtar.
+            AllBooksBtn = (Button) findViewById(R.id.all_books);
+            AllBooksBtn.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    openAllBooksActivity();
+                }
+            });
 
         }
 
@@ -128,6 +122,10 @@ public class MainActivity extends AppCompatActivity {
     }
     public void openRequestBookActivity() {
         Intent intent = new Intent(this, RequestBookActivity.class);
+        startActivity(intent);
+    }
+    public void openAllBooksActivity() {
+        Intent intent = new Intent(this, AllBooksActivity.class);
         startActivity(intent);
     }
 }
